fix(script): code block escaping and hljs
